separated rendering functions and sidebar functions to different files; use index.js for exports only; added new rule to eslint to remove parentheses; added blocking grid unit test
